+++ tests/tests/models/test_file.py
@@ -5,6 +5,7 @@ from paper_uploads.models import UploadedFile
 
 from ..dummy import *
 from .test_dummy import (
+    BacklinkModelMixin,
     TestFileFieldResource,
     TestFileFieldResourceAttach,
     TestFileFieldResourceDelete,
@@ -13,7 +14,7 @@ from .test_dummy import (
 )
 
 
-class TestUploadedFile(TestFileFieldResource):
+class TestUploadedFile(BacklinkModelMixin, TestFileFieldResource):
     resource_url = '/media/files/%Y-%m-%d'
     resource_location = 'files/%Y-%m-%d'
     resource_name = 'Nature Tree'
@@ -23,7 +24,7 @@ class TestUploadedFile(TestFileFieldResource):
     owner_app_label = 'app'
     owner_model_name = 'fileexample'
     owner_fieldname = 'file'
-    owner_class = FileExample
+    owner_model = FileExample
     file_field_name = 'file'
 
     @classmethod
@@ -70,12 +71,13 @@ class TestUploadedFileAttach(TestFileFieldResourceAttach):
     resource_class = UploadedFile
 
 
-class TestUploadedFileRename(TestFileFieldResourceRename):
+class TestUploadedFileRename(BacklinkModelMixin, TestFileFieldResourceRename):
     resource_class = UploadedFile
     resource_location = 'files/%Y-%m-%d'
     owner_app_label = 'app'
     owner_model_name = 'fileexample'
     owner_fieldname = 'file'
+    owner_model = FileExample
 
     @classmethod
     def init_class(cls, storage):
@@ -100,12 +102,13 @@ class TestUploadedFileRename(TestFileFieldResourceRename):
         storage.resource.delete()
 
 
-class TestUploadedFileDelete(TestFileFieldResourceDelete):
+class TestUploadedFileDelete(BacklinkModelMixin, TestFileFieldResourceDelete):
     resource_class = UploadedFile
     resource_location = 'files/%Y-%m-%d'
     owner_app_label = 'app'
     owner_model_name = 'fileexample'
     owner_fieldname = 'file'
+    owner_model = FileExample
 
     @classmethod
     def init_class(cls, storage):
@@ -132,13 +135,18 @@ class TestUploadedFileEmpty(TestFileFieldResourceEmpty):
     recource_class = UploadedFile
 
 
-class TestUploadedFileExists:
-    @staticmethod
-    def init_class(storage):
+class TestUploadedFileExists(BacklinkModelMixin):
+    owner_app_label = 'app'
+    owner_model_name = 'fileexample'
+    owner_fieldname = 'file'
+    owner_model = FileExample
+
+    @classmethod
+    def init_class(cls, storage):
         storage.resource = UploadedFile(
-            owner_app_label='app',
-            owner_model_name='fileexample',
-            owner_fieldname='file'
+            owner_app_label=cls.owner_app_label,
+            owner_model_name=cls.owner_model_name,
+            owner_fieldname=cls.owner_fieldname
         )
         with open(NATURE_FILEPATH, 'rb') as fp:
             storage.resource.attach_file(fp)